requirements.txt
torch==1.12.0+cu116
torchvision==0.13.0+cu116
timm==1.0.25

opencv-python-headless==4.11.0.86
pillow==11.3.0
scikit-image==0.25.2
imageio==2.37.3
tifffile==2025.5.10

numpy==1.26.4
scipy==1.15.3
pandas==2.3.2

einops==0.8.2
tensorboardX==2.6.4
lpips==0.1.4
clean-fid==0.1.35
pytorch-fid==0.3.0
